rrr: convert misc/any_message to module (rrr.any_message)
any_message.cpp becomes the rrr.any_message module interface unit;
any_message.hpp becomes a one-line shim that imports it. Exports:
  - IdempotencyKey is misnamed (actually this file): AnyMessage class
    with save/load + type-erased payload (SerializableProxy)
  - AnyMessageRegistry: name -> factory registry + reverse
    type_index -> name map (rusty::HashMap behind SpinMutex)
  - reg_any_message_as<T>(name) template helper
  - Inline AnyMessage::is_a<T> / unpack<T> / pack_as<T> / pack<T>
    templates
  - operator<<(BinaryWriteArchive&, AnyMessage&) +
    operator>>(BinaryReadArchive&, AnyMessage&)

Imports: rrr.debugging (verify), rrr.serializable (BinaryWriteArchive,
SerializableProxy, SerializableSharedPtrHolder), rrr.threading
(SpinMutex). The textual `#include "../misc/marshal.hpp"` that used
to bring Marshal in transitively is gone — any_message doesn't need
Marshal directly.

30 modules total.

Measurement (rrr + rpcbench, -j32, clean):
  +marshal:        61.70 s, 350 MB PCM
  +any_message:    56.85 s, ~351 MB PCM

Wallclock dipped 5 s — well within run-to-run noise (the underlying
parallelism shape depends on critical-path BMI build order).
