+++ src/rrr/CMakeLists.txt
@@ -25,6 +25,7 @@ set(RRR_MODULE_SRC
     ${CMAKE_CURRENT_SOURCE_DIR}/base/strop.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/base/threading.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/misc/alarm.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/misc/any_message.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/misc/cpuinfo.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/misc/dball.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/misc/marshal.cpp
@@ -79,6 +80,7 @@ if(NOT TARGET rrr)
         # base/debugging.cpp omitted — module unit.
         ${CMAKE_CURRENT_SOURCE_DIR}/base/threading.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/misc/alock.cpp
+        # misc/any_message.cpp omitted — module unit.
         # misc/marshal.cpp omitted — module unit.
         # base/strop.cpp omitted — converted to a C++23 module interface
         # unit (see RRR_MODULE_SRC). rusty-cpp cannot resolve module
